Move functional test plotting functions to new module.
Allows unit tests to be able to run without requiring matplotlib.
