+++ testing/functional_testing/patch/patch_test.py
@@ -6,7 +6,8 @@ import xarray as xr
 import pandas as pd
 import numpy as np
 import matplotlib.pyplot as plt
-from utils import round_up, get_color_palette, blank_plot
+from utils import round_up
+from utils_plotting import blank_plot, get_color_palette
 from functional_class import FunctionalTest
